Fix Kamino pointer offsets beyond two GiB
Promote element offsets before computing byte addresses in the RCM
pointer helper. Multiplying in int32 first wraps large Schur and
factorization buffer offsets on CUDA.

Cover float32 and int32 offsets across the two- and four-GiB boundaries
without allocating large buffers. The CUDA regression fails before the
fix; all 14 pointer and RCM tests pass afterward.

+++ newton/_src/solvers/kamino/_src/linalg/factorize/llt_blocked_rcm.py
@@ -85,7 +85,7 @@ def make_get_array_offset_ptr_func(dtype):
 
     @wp.func
     def get_dtype_array_offset_ptr(arr: wp.array[dtype], start_index: int) -> wp.uint64:
-        return get_dtype_array_ptr(arr) + wp.uint64(start_index * wp.static(sizeof(dtype._type_)))
+        return get_dtype_array_ptr(arr) + wp.uint64(start_index) * wp.uint64(wp.static(sizeof(dtype._type_)))
 
     return get_dtype_array_offset_ptr
